[refactor] dummy Chart 리팩토링
- dummy chart 코드를 `issue.tsx`에서 `components/Chart`로 분리
